+++ FE/src/components/header/index.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 
 import GNB from 'components/header/gnb';
 import * as Styled from 'components/header/header.style';
+import { ModalContext } from 'components/context/ModalContext';
 import SearchBar from './searchBar';
 
 function Header() {
@@ -9,7 +10,9 @@ function Header() {
     <Styled.Header>
       <Styled.HeaderWrapper>
         <GNB />
-        <SearchBar />
+        <ModalContext>
+          <SearchBar />
+        </ModalContext>
       </Styled.HeaderWrapper>
     </Styled.Header>
   );
